fix bug, refactor ColorField
