+++ translations/ja.yml
@@ -174,7 +174,11 @@ src:
             ColorField:
               '1115770113': 保存
               '1533357798': 未設定
+              '1847234945': 赤
+              '1936615876': アルファ
+              '2090560987': 緑
               '3528672691': キャンセル
+              '3801160378': 青
             LocationField:
               '2222144661': 緯度
               '3067266584': 高度
